#5 added support for bean validation

+++ neo4j/src/main/java/com/buschmais/cdo/neo4j/impl/node/proxy/method/object/HashCodeMethod.java
@@ -7,6 +7,6 @@ public class HashCodeMethod implements NodeProxyMethod {
 
     @Override
     public Object invoke(Node node, Object instance, Object[] args) {
-        return Long.valueOf(node.getId());
+        return Integer.valueOf((int) node.getId());
     }
 }
